Added Application config file
