+++ main/midnight/Classes/panels/panel_splashscreen.cpp
@@ -5,7 +5,7 @@
 #include "../system/resolutionmanager.h"
 #include "../ui/uihelper.h"
 #include "../system/moonring.h"
-#include "../system/configmanager.h"
+#include "../system/settingsmanager.h"
 #include "../system/panelmanager.h"
 #include "../system/progressmonitor.h"
